ledger-on-memory: Split the tests into a test suite. (#7069)
* ledger-on-memory: Split the tests into a test suite.

The ledger-on-memory tests are sporadically failing on CI because they
hit the timeout of 60s, which is the default for a "small" test target.
This is not outside the realm of possibility, as tests can take a while
on CI when lots of other things are running.

To mitigate this, rather than increasing the timeout, I am splitting the
tests up into a suite, which means they'll run as 3 targets, not 1, each
with a 60s timeout. This means they're far more likely to complete in
the allocated time.

CHANGELOG_BEGIN
CHANGELOG_END

* ledger-on-memory: Don't implicitly widen numerics.

Not sure why this breaks on CI and not on my machine, but 🤷🏽‍♂️.