let users stop listening to event streams
I have to imagine this is a typo: there currently is no way to
unsubscribe from a stream without closing it, as the `off` method we
expose actually redirects to an `on` method internally, which means that
instead of the method no longer being called, it now gets called twice
(or rather, one more time) on each event.

That seems a bit rude.

CHANGELOG_BEGIN

- [JavaScript Client Libraries] Bugfix: calling the `off` method of
  event streams returned by `streamQuery` and `streamFetchByKey` now
  correctly removes the given listener, rather than adding it again.

CHANGELOG_END