kvutils: Move execution contexts to the submission methods. (#7071)
* kvutils: Move execution contexts to the submission methods.

Previously, the execution contexts were provided at the constructor
level, which meant the submission validator and associated components
would use the resource acquisition execution context. This context is
intended purely for resource acquisition; if we need an execution
context for processing data, we have much more control when we
explicitly create one.

Implementors will still need to avoid passing an execution context
implicitly into the constructors of their own components, instead
favoring an explicit one that is to be used during `commit` or reading
events.

CHANGELOG_BEGIN
- [Integration Kit] kvutils now expects execution contexts to be passed
  in to the various ``SubmissionValidator``, ``LedgerStateAccess``, and
  ``LedgerStateOperations`` methods. This is a source-breaking change.

  Instead of providing an execution context implicitly to your ledger
  implementation, you are encouraged to construct the relevant contexts
  for different operations explicitly. Please refer to the open-source
  implementations as a reference.
CHANGELOG_END

* ledger-on-memory: Don't use the acquisition context for commits.

Instead, use the materializer execution context; it may not be the best
choice but at least it's appropriate.

* kvutils: Standardize method parameter formatting.

* kvutils: Provide old state  APIs as `LedgerStateAccess.v1_4`.