increase Maven timeout
We reach the timeout on pretty much every release lately and we need to
retry to entire release process two or three times to get through. This
is us giving up on Maven, not Maven itself crashing, so I4d like to try
giving it more time.

CHANGELOG_BEGIN
CHANGELOG_END