kvutils: Sort the output state before writing. [KVL-558] (#7515)
* kvutils: Always commit serially.

Committing in parallel will cause the set of writes to be emitted in a
random order.

CHANGELOG_BEGIN
- [Integration Kit] In kvutils, the ``BatchedSubmissionValidator`` no
  longer has a parameter for commit parallelism. Commits are now always
  written serially to preserve order.
CHANGELOG_END

* kvutils: Sort the output state before writing.

We previously wrote the output state in a random order, leading to
writes that could not be easily compared for integrity. Sorting them
allows us to validate not just the values, but the order.

This means that the exporter no longer sorts the write set, and the
integrity checker does not sort before checking.

CHANGELOG_BEGIN
- [Integration Kit] In kvutils, state is now sorted before committing.
  This allows us to provide stronger guarantees with regards to the
  serialized write sets.

  If you have implemented your own ``CommitStrategy``, you should also
  ensure the output state is sorted before committing.
CHANGELOG_END

* ledger-on-sql: Implement SQL typeclasses as values, not functions.

Let's pretend performance is important here.

* kvutils: Allow Iterables in batching ledger state operations.

A `Map[Key, Value]` is an `Iterable[(Key, Value)]`, but not a `Seq`.
This allows us to pass a `Map` straight from the committer.

* kvutils: Construct a `SortedMap` in the commit strategy.

This should be more performant than generating and then sorting a
`Vector`.

* kvutils: Sort state updates when committing after pre-execution.