Speedy Compiler: factorize and clean code (fixe of #7493)  (#7540)
* Revert "Revert "Speedy Compiler: factorize and clean code (#7493)""

This reverts commit d99f35b3cd98a666af46fada3a75c8d2a194964c.

* Speedy: Reintroduce usage of muti-binding let.

This PR fixes recursion issue with `closureConvert` and `freeVars`.
The explosion of muti-binding in single binding makes the function
`closureConvert` crash with stackoverflow in some cases.

We reintroduce the usage of multi-binding let in the first pass of the
compiler. The ANF transformation pass will replace them with nested
single binding let. ANF pass do not have issue with deep expression,
as it handle them with trampoline.

CHANGELOG_BEGIN
CHANGELOG_END