react: factor out stream handling (#7559)
A user has [asked on the forum] how their application could handle
WebSocket errors. The unfortunate response at the moment is that they
can't: our React bindings will just spit out an error log, and that's
it. The current code seems to just assume the stream will eventually
reconnect, but the underlying stream does stop trying if it fails two
times in a row faster than `reconnectThreshold`.

[asked on the forum]: https://discuss.daml.com/t/usestreamquery-disconnecting/1325

In trying to address this, and with the context of at some point
expanding the React bindings to the multi-{key,query} API, I realized
that, given the current structure of the code, I might end up having to
solve this issue four times, so I decided to first factor out the stream
handling logic in the existing code, and I believe this makes sense as a
separate PR.

I should note, however, that this is _not_ a refactoring: as indicated
by the amended tests, this PR actually changes the behaviour of
`useStreamFetchByKey`. I believe this counts as a bugfix, but welcome
any pointer as to why the behaviours of `useStreamQuery` and
`useStreamFetchByKey` should differ with respect to the `live` event.

CHANGELOG_BEGIN
  * JavaScript Client Libraries: fix a bug where the `useStreamFetchByKey`
    hook would, in some circumstances, report a "ready" state (i.e.
    `loading: false`) even though the underlying connection had not yet been
    fully established.
CHANGELOG_END