Bump hardcoded timeout for party alloc/package upload
This is a short term fix to remediate issues with uploading packages
that take a considerable amount of time to decode and validate and
therefore exhausting the 30 seconds.

Adding a maximum record parameter to the ledger API like we already have
for the config management service is not as straight forward for the
package upload, because one has to account for the time in transit as
well. This topic needs further analysis, but in the meantime raising the
default to 2 minutes should provide enough headroom to alleviate
existing issues with package upload timing.

Contributes to #6880

CHANGELOG_BEGIN
[Integration Kit, Ledger API Server]: The hardcoded timeout for party
allocation and package uploads has been increased to 2 minutes. See
`issue #6880 <https://github.com/digital-asset/daml/issues/6880>`__.
CHANGELOG_END