LF: enforce non-empty maintainers in contract key in fetchByKey
This fixes a bug in the Speedy interpreter. With this change, the
interpreter crashes if it attempts to fetch using a contract key that
has an empty set of maintainers. This propagates to exerciseByKey as
this command is compiled using fetchByKey speedy builtin.

This is a breaking change approved by @bame-da .

CHANGELOG_BEGIN

    [LF] (Bug fix) enforce non-empty maintainers in contract key during
        exericiseByKey.

CHANGELOG_END