Warn on potentially un-upgradable language extensions
We issue a warning whenever a module is compiled with a language
extension enabled for which we are not certain that it works with
`data-dependencies`. Currently, we consider all extensions except for
the ones enables by default, `ApplicativeDo` and
`PartialTypeSignatures` problematic in this regard. The list of
extensions that works fine with `data-dependencies` but we need to do
further research and add further tests before we add any extensions.

The warning is always put on the first line of the file regardless of
where the `{-# LANGUAGE ... #-}` pragma is actually used. This is due
to the fact that the `ParsedModule` we get from GHC does not contain
the location information of these pragmas. We should probably improve
this over time but I think it is acceptable for now.

We also only issue the warning in DAML Studio right now. I'll add
support for `daml build` in a separate PR.

CHANGELOG_BEGIN
[DAML Compiler] Warn when a module uses a language extension that
might not work with `data-dependencies`.
CHANGELOG_END