LF: validate Node with byKey flag.
This adapate and test the validation of transaction containing nodes
with `byKey` flag.

With respect to #7617, the type of `byKey` is changed from Boolean to
Option[Boolean]. The idea is that for old transactions the flag will
be set to None (because we don't know if the node was actually done
"by key" or not) while for the next transaction version the flag will
be set to `Some(true)` or `Some(false)` (to be done in an upcoming
PR). This allows us to validate old transaction in a backward
compatible way without injecting the serialization version in the node
comparison.

This advances the state of #7622

CHANGELOG_BEGIN
CHANGELOG_END